Moved components from utils.py and test_utils.py to other files.
